backend/pyproject.toml
[project]
name = "planet-ops-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.11",
    "alembic>=1.18.4",
    "anthropic>=0.84.0",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "click>=8.1.7",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "psutil>=7.2.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.2",
    "slack-sdk>=3.37.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
commander = "app.cli:cli"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
